Create R&D_Report_02
I was given a invoice PDF file from Robi. I was asked to extract a certain part of the file using python. 

After researching on it. I’ve found that it can be easily done with the Tabula Web Tool. It generates a CVS file. Alternatively, I’ve come up with the following code to save the data to an excel sheet: